renamed user-controller to users-controller
