+++ controllers/index.js
@@ -1,4 +1,4 @@
-const UsersController = require('./user-controller');
+const UsersController = require('./users-controller');
 
 
 module.exports = {
